--- project/src/components/not-found-page/not-found-page.tsx
@@ -1,5 +1,0 @@
-function NotFoundPage(): JSX.Element {
-  return <h1>404. Page not found</h1>;
-}
-
-export default NotFoundPage;
