+++ project/src/components/private-route/private-route.tsx
@@ -1,0 +1,13 @@
+import {Navigate} from 'react-router-dom';
+
+type PrivateRouteProps = {
+  children: JSX.Element;
+};
+
+function PrivateRoute({children}: PrivateRouteProps): JSX.Element {
+  const hasAccess = false;
+
+  return hasAccess ? children : <Navigate to={'/login'} />;
+}
+
+export default PrivateRoute;
