+++ project/src/pages/not-found-page/not-found-page.tsx
@@ -1,0 +1,10 @@
+function NotFoundPage(): JSX.Element {
+  return (
+    <div style={{ marginTop: '40vh', textAlign: 'center' }}>
+      <h1>404. Page not found</h1>
+      <a href="/">Вернуться на главную</a>
+    </div>
+  );
+}
+
+export default NotFoundPage;
